// app/types.ts
export interface Legislator {
  cid: string;
  firstlast: string;
  lastname: string;
  party: string;
  office: string;
  gender: string;
  first_elected: string;
  exit_code: string;
  comments: string;
  phone: string;
  fax: string;
  website: string;
  webform: string;
  congress_office: string;
  bioguide_id: string;
  votesmart_id: string;
  feccandid: string;
  twitter_id: string;
  youtube_url: string;
  facebook_id: string;
  birthdate: string;
}